Add files via upload
This is a brief project that I started that involves experimenting with the human eye. The goal is to produce function that can generate a stereogram for an arbitrary image. Some of the parameters of the function include an  image in grayscale format along with hshift (number of pixels etched part is shifted horizontally in the dispatch)  which can be calculated /approximated depending on the parallax.